Move GetBrokerConfig() to the api
