+++ pkg/resources/kafka/kafka.go
@@ -206,7 +206,7 @@ func (r *Reconciler) Reconcile(log logr.Logger) error {
 
 	brokersVolumes := make(map[string][]*corev1.PersistentVolumeClaim, len(r.KafkaCluster.Spec.Brokers))
 	for _, broker := range r.KafkaCluster.Spec.Brokers {
-		brokerConfig, err := util.GetBrokerConfig(broker, r.KafkaCluster.Spec)
+		brokerConfig, err := v1beta1.GetBrokerConfig(broker, r.KafkaCluster.Spec)
 		if err != nil {
 			return errors.WrapIf(err, "failed to reconcile resource")
 		}
@@ -229,7 +229,7 @@ func (r *Reconciler) Reconcile(log logr.Logger) error {
 
 	reorderedBrokers := r.reorderBrokers(log, r.KafkaCluster.Spec.Brokers)
 	for _, broker := range reorderedBrokers {
-		brokerConfig, err := util.GetBrokerConfig(broker, r.KafkaCluster.Spec)
+		brokerConfig, err := v1beta1.GetBrokerConfig(broker, r.KafkaCluster.Spec)
 		if err != nil {
 			return errors.WrapIf(err, "failed to reconcile resource")
 		}
@@ -879,7 +879,7 @@ func (r *Reconciler) createExternalListenerStatuses(log logr.Logger) (map[string
 				portNumber := eListener.ExternalStartingPort + broker.Id
 
 				if eListener.GetAccessMethod() != corev1.ServiceTypeLoadBalancer {
-					bConfig, err := util.GetBrokerConfig(broker, r.KafkaCluster.Spec)
+					bConfig, err := v1beta1.GetBrokerConfig(broker, r.KafkaCluster.Spec)
 					if err != nil {
 						return nil, err
 					}
@@ -902,7 +902,7 @@ func (r *Reconciler) createExternalListenerStatuses(log logr.Logger) (map[string
 						brokerHost = fmt.Sprintf("%s-%d-%s.%s%s", r.KafkaCluster.Name, broker.Id, eListener.Name, r.KafkaCluster.Namespace, brokerHost)
 					}
 				}
-				brokerConfig, err := util.GetBrokerConfig(broker, r.KafkaCluster.Spec)
+				brokerConfig, err := v1beta1.GetBrokerConfig(broker, r.KafkaCluster.Spec)
 				if err != nil {
 					return nil, err
 				}
